+++ .github/CODEOWNERS
@@ -3,5 +3,5 @@
 # https://github.blog/2017-07-06-introducing-code-owners/
 
 # Add a fallback to default owners
-*       @thebino
+*       @thebino @crocsandcoffee
