+++ settings.gradle.kts
@@ -8,13 +8,14 @@ include(":app-desktop")
 //include(":app-iOS")
 
 include(":ui:timeline")
-include(":ui:welcome")
 //include(":ui:messages")
 //include(":ui:notifications")
 //include(":ui:search")
 //include(":ui:settings")
 include(":ui:common")
-include(":ui:splash")
+include(":ui:root")
+include(":ui:signed-in")
+include(":ui:signed-out")
 
 include(":domain:timeline")
 include(":domain:authentication")
