+++ ui/root/src/androidMain/AndroidManifest.xml
